Attempt to separate out MLIRClang library
