+++ mlir/tools/mlir-clang/Lib/CMakeLists.txt
@@ -1,0 +1,40 @@
+
+add_mlir_library(MLIRClang
+  clang-mlir.cc
+
+  EXCLUDE_FROM_LIBMLIR
+
+  DEPENDS
+  mlir-headers
+  ClangDriverOptions
+
+  LINK_COMPONENTS
+  Core
+  Support
+
+  LINK_LIBS PUBLIC
+
+  MLIRSCFTransforms
+
+  MLIRSupport
+  MLIRIR
+  MLIRAnalysis
+  MLIRLLVMIR
+  MLIRNVVMIR
+  MLIRGPU
+  MLIRTransforms
+  MLIRSCFToStandard
+  MLIRStandardToLLVM
+  MLIRAffineTransforms
+  MLIRAffineToStandard
+  MLIRTargetLLVMIRImport
+
+  clangAST
+  clangBasic
+  clangCodeGen
+  clangDriver
+  clangFrontend
+  clangFrontendTool
+  clangLex
+  clangSerialization
+)
